detector/requirements.txt
flask
psutil
pyyaml
requests